Squash 2 skin files
