--- includes/SkinTGUI22.php
@@ -1,104 +1,0 @@
-<?php
-
-namespace MediaWiki\Skins\TGUI;
-
-/**
- * @ingroup Skins
- * @package TGUI
- * @internal
- */
-class SkinTGUI22 extends SkinTGUI {
-	/**
-	 * Determines if the Table of Contents should be visible.
-	 * TOC is visible on main namespaces except for the Main Page.
-	 *
-	 * @internal
-	 * @return bool
-	 */
-	public function isTableOfContentsVisibleInSidebar(): bool {
-		$title = $this->getTitle();
-
-		if (
-			!$title ||
-			$title->isMainPage()
-		) {
-			return false;
-		}
-
-		if ( $this->isTOCABTestEnabled() ) {
-			return $title->getArticleID() !== 0;
-		}
-
-		return true;
-	}
-
-	/**
-	 * Annotates table of contents data with TGUI-specific information.
-	 *
-	 * In tableOfContents.js we have tableOfContents::getTableOfContentsSectionsData(),
-	 * that yields the same result as this function, please make sure to keep them in sync.
-	 *
-	 * @param array $tocData
-	 * @return array
-	 */
-	private function getTocData( array $tocData ): array {
-		// If the table of contents has no items, we won't output it.
-		// empty array is interpreted by Mustache as falsey.
-		if ( empty( $tocData ) || empty( $tocData[ 'array-sections' ] ) ) {
-			return [];
-		}
-
-		// Populate button labels for collapsible TOC sections
-		foreach ( $tocData[ 'array-sections' ] as &$section ) {
-			if ( $section['is-top-level-section'] && $section['is-parent-section'] ) {
-				$section['tgui-button-label'] =
-					$this->msg( 'tgui-toc-toggle-button-label', $section['line'] )->text();
-			}
-		}
-
-		return array_merge( $tocData, [
-			'is-tgui-toc-beginning-enabled' => $this->getConfig()->get(
-				'TGUITableOfContentsBeginning'
-			),
-			'tgui-is-collapse-sections-enabled' =>
-				$tocData[ 'number-section-count'] >= $this->getConfig()->get(
-					'TGUITableOfContentsCollapseAtCount'
-				)
-		] );
-	}
-
-	/**
-	 * Merges the `view-overflow` menu into the `action` menu.
-	 * This ensures that the previous state of the menu e.g. emptyPortlet class
-	 * is preserved.
-	 * @param array $data
-	 * @return array
-	 */
-	private function mergeViewOverflowIntoActions( $data ) {
-		$portlets = $data['data-portlets'];
-		$actions = $portlets['data-actions'];
-		$overflow = $portlets['data-views-overflow'];
-		// if the views overflow menu is not empty, then signal that the more menu despite
-		// being initially empty now has collapsible items.
-		if ( !$overflow['is-empty'] ) {
-			$data['data-portlets']['data-actions']['class'] .= ' tgui-has-collapsible-items';
-		}
-		$data['data-portlets']['data-actions']['html-items'] = $overflow['html-items'] . $actions['html-items'];
-		return $data;
-	}
-
-	/**
-	 * @return array
-	 */
-	public function getTemplateData(): array {
-		$featureManager = TGUIServices::getFeatureManager();
-		$parentData = parent::getTemplateData();
-
-		$parentData = $this->mergeViewOverflowIntoActions( $parentData );
-
-		return array_merge( $parentData, [
-			// Cast empty string to null
-			'html-subtitle' => $parentData['html-subtitle'] === '' ? null : $parentData['html-subtitle'],
-		] );
-	}
-}
