+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "TGUI",
   "private": true,
-  "version": "0.1.0-mediawiki",
+  "version": "0.2.0-mediawiki",
   "scripts": {
     "test": "npm -s run lint && tsc && npm run test:unit",
     "test:unit": "jest --silent",
