+++ resources/skins.tgui.styles/components/MenuDropdown.less
@@ -31,7 +31,7 @@
     padding: 0 0.5em;
     margin-bottom: 0.5em;
     background-color: var(--color-primary);
-    color: var(--color-button-text);
+    color: var(--color-button-text) !important;
     border-radius: var(--border-radius-small);
     transition: background-color var(--transition-time);
     text-decoration: none;
